Remove unnecessary copies of ca.h and cert.h
